Tests/LibWeb/Text/expected/HTML/img-pending-load-renders-blank-not-alt-text.txt

pending image renders blank: false
broken image renders alt text: true
src-less image renders alt text: true
deferred lazy image renders blank: false
unblocked image loaded with natural width: 1
